feat: Add `status` and `facility` options to `netbox_location` module (#1282)
* feat: Add `status` and `facility` options to `netbox_location` module

This commit adds the `status` and `facility` options to the `netbox_location` module. The `status` option allows specifying the status of the location, while the `facility` option allows specifying the data center provider or facility. These options were added to enhance the functionality of the module and provide more flexibility in managing locations within NetBox.

Fixes #1279, #1280

* remove extra space found in linting