feat: Include netbox_tunnel.yml in integration tests
This commit adds the `netbox_tunnel.yml` file to the integration tests for targets v3.5, v3.6, v3.7, and v4.0. The file is included using the `ansible.builtin.include_tasks` module with the `netbox_tunnel.yml` file as the source. This change ensures that the `netbox_tunnel` tests are executed as part of the integration test suite.